Revert "change made for button to be in seperate component"
This reverts commit 3e48006eed0a7029d7ec38a48227c81c850688e8.
