+++ app/src/stylesheets/App.css
@@ -32,11 +32,6 @@ body {
 }
 
 /* Todo Button */
-.todo-btn {
-  position: absolute;
-  border: 2px solid red;
-}
-
 .circle {
   background-color: dodgerblue;
   height: 60px;
